fix: Standardize IMU units to SI and remove deprecated capabilities
Makes all IMU source and target devices convert to/from standard SI units.

Removes the Gamepad::Accelerometer and Gamepad::Gyro capabilities and
all workarounds related to them.

Also corrects the Steam Deck's accel/gyro scale constants, which were
off by about 2.4% from SDL's real +/-2g / +/-2000 deg/s range.

+++ src/drivers/ultimate_2/event.rs
@@ -86,9 +86,9 @@ pub enum TriggerEvent {
 /// [InertialInput] represents the state of the IMU (x, y, z) values
 #[derive(Clone, Debug)]
 pub struct InertialInput {
-    pub x: i16,
-    pub y: i16,
-    pub z: i16,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 /// [InertialEvent] has data from the IMU
